+++ src/Fabulous.MauiControls.Tests/Library.fs
@@ -1,4 +1,0 @@
-﻿namespace Fabulous.MauiControls.Tests
-
-module Say =
-    let hello name = printfn "Hello %s" name