+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix the crash at startup when targeting Windows by using FSharp.Maui.WinUICompat by @TimLariviere (https://github.com/fabulous-dev/Fabulous.MauiControls/pull/10)
 - Fix an issue where template could enumerate the whole disk to add under the Android target folder by @Smaug123 (https://github.com/fabulous-dev/Fabulous.MauiControls/pull/9)
+- Write a custom NavigationPage to use synchronous push/pop and call Mounted/Unmounted events on pages by @TimLariviere (https://github.com/fabulous-dev/Fabulous.MauiControls/pull/13)
 
 ## [2.2.0] - 2023-01-24